fix(critical): 강제진입 시 포지션 side 매핑 로직 수정
업비트 buy = 숏 포지션 (김프 축소 시 이익)
업비트 sell = 롱 포지션 (김프 확대 시 이익)

기존에는 반대로 매핑되어 손익 계산이 잘못됨
포지션 1494도 long → short로 DB 수정 완료

+++ server/services/unified-position.ts
@@ -98,7 +98,8 @@ export class UnifiedPositionService {
         strategyId: params.strategyId,
         symbol: params.symbol,
         type: "kimchi_arbitrage",
-        side: (params.side === 'buy' ? 'long' : params.side === 'sell' ? 'short' : (params.side as 'long' | 'short')),
+        // 김치 프리미엄 거래: 업비트 buy = 숏 포지션, 업비트 sell = 롱 포지션
+        side: (params.side === 'buy' ? 'short' : params.side === 'sell' ? 'long' : (params.side as 'long' | 'short')),
         status: "open" as const,
         entryPrice: safePrice,
         quantity: safeQty,
